[#1] KotlinJVM 분리
