+++ build-logic/convention/src/main/java/team/noweekend/convention/KotlinAndroid.kt
@@ -3,7 +3,6 @@ package team.noweekend.convention
 import com.android.build.api.dsl.CommonExtension
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.kotlin.dsl.assign
 import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.provideDelegate
@@ -17,32 +16,20 @@ internal fun Project.configureKotlinAndroid(
 ) {
     commonExtension.apply {
         compileSdk = 35
-        
+
         defaultConfig {
             minSdk = 26
         }
-        
+
         compileOptions {
             sourceCompatibility = JavaVersion.VERSION_17
             targetCompatibility = JavaVersion.VERSION_17
             isCoreLibraryDesugaringEnabled = false
         }
     }
-    
+
     configureKotlin<KotlinAndroidProjectExtension>()
 }
-
-internal fun Project.configureKotlinJvm() {
-    extensions.configure<JavaPluginExtension> {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-    
-    extensions.configure<KotlinJvmProjectExtension> {
-        jvmToolchain(jdkVersion = 17)
-    }
-}
-
 
 private inline fun <reified T : KotlinTopLevelExtension> Project.configureKotlin() = configure<T> {
     // Treat all Kotlin warnings as errors (disabled by default)
