+++ build-logic/convention/src/main/java/team/noweekend/convention/KotlinJVM.kt
@@ -1,0 +1,18 @@
+package team.noweekend.convention
+
+import org.gradle.api.JavaVersion
+import org.gradle.api.Project
+import org.gradle.api.plugins.JavaPluginExtension
+import org.gradle.kotlin.dsl.configure
+import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
+
+internal fun Project.configureKotlinJvm() {
+    extensions.configure<JavaPluginExtension> {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+
+    extensions.configure<KotlinJvmProjectExtension> {
+        jvmToolchain(jdkVersion = 17)
+    }
+}
